Added option to full build source directory
