+++ content/Embedded Example.html
@@ -7,12 +7,14 @@
         </title>
     </head>
     <body>
-        <article>
+<article>
     <header>Created: 2025-01-11 <br/>Last Modified: 2025-01-11</header>
-    This is an example of embedded content:
-<br />
-![[example#Now onto some code]]
+		<section class=h2><h2 id="Content-from-Example">Content from Example</h2><br />
+			This is an example of embedded content:<br />
+			<br /><br />
+			![[example#Now onto some code]]<br />
+			<br /><br />
+		</section><br />
 
-</article>
-    </body>
+</article>    </body>
 </html>
