+++ content/Nested Example.html
@@ -1,0 +1,46 @@
+<!DOCTYPE html>
+<html lang="en">
+    <link rel="stylesheet" href="python.css" />
+    <head>
+        <title>
+            Nested Example
+        </title>
+    </head>
+    <body>
+<article>
+    <header>Created: 2025-01-12 <br/>Last Modified: 2025-01-12</header>
+		<section class=h2><h2 id="This-is-H2">This is H2</h2><br />
+			Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras ullamcorper eros at maximus porta.<br />
+			<br /><br />
+			<section class=h3><h3 id="This-is-H3">This is H3</h3><br />
+				Vestibulum lectus metus, ornare nec suscipit at, sagittis vitae augue. Vestibulum sem velit, lobortis eu mattis nec, pulvinar ac metus.<br />
+				<br /><br />
+				<section class=h4><h4 id="This-is-H4">This is H4</h4><br />
+					Nulla ex quam, gravida in efficitur vel, accumsan vitae lorem. In at malesuada mauris. Etiam id dui sit amet nisl efficitur fermentum.<br />
+					<br /><br />
+					<section class=h5><h5 id="This-is-H5">This is H5</h5><br />
+						Nam mattis sapien eget dolor imperdiet suscipit. Phasellus tristique eleifend dolor nec ultrices. Fusce semper ex mollis risus vehicula fringilla. Vivamus commodo ut tortor vitae ultricies. Mauris consequat fermentum egestas. Nunc pulvinar mauris ipsum, a ornare massa consectetur eget.<br />
+						<br /><br />
+						<section class=h6><h6 id="This-is-H6">This is H6</h6><br />
+							Donec venenatis rhoncus porttitor. Aliquam sed luctus mi. Fusce vel massa vel magna luctus sollicitudin ut vitae magna. Donec pulvinar consequat nisi, convallis elementum metus condimentum id. Mauris gravida eros non odio varius ornare. Mauris vel orci tincidunt, lobortis augue nec, ultrices lacus. Etiam feugiat sagittis rutrum. Fusce in sapien aliquam, pellentesque risus eu, iaculis est. Nam convallis, mi eget maximus sagittis, justo nisi elementum ex, ut luctus risus mi eu ipsum. Aliquam sit amet egestas diam. Donec non massa ac turpis tempus suscipit.<br />
+							<br /><br />
+						</section><br />
+						<section class=h6><h6 id="This-is-Another-H6">This is Another H6</h6><br />
+							Etiam ultricies massa et placerat placerat. Sed tempus nec metus vel ullamcorper. Vestibulum tempus id mauris in eleifend. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris suscipit tincidunt tellus, in auctor eros laoreet et. Vivamus nisl ligula, facilisis non interdum accumsan, dapibus vel augue. Ut aliquam ligula a dolor lobortis pellentesque. Vivamus nec neque volutpat massa dignissim aliquam sed nec nibh. Nullam vitae ullamcorper est. Morbi condimentum dolor non nisl hendrerit fringilla.<br />
+							<br /><br />
+						</section><br />
+					</section><br />
+				</section><br />
+			</section><br />
+			<section class=h3><h3 id="Here's-an-H3">Here's an H3</h3><br />
+				Quisque tempus ipsum eros, non vulputate mauris venenatis eu. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed vehicula magna sit amet leo fringilla, non pellentesque ante porta. Integer semper dapibus augue nec accumsan. Duis felis turpis, auctor sit amet lacinia et, tincidunt ac mi. Donec elementum mauris et dictum suscipit. Duis pellentesque augue lectus. Praesent gravida molestie interdum. Vivamus scelerisque erat vel diam efficitur, sit amet mollis urna dapibus. Proin sed fringilla est. Nulla ultricies nisl ut orci eleifend semper. Aliquam eu nulla ligula. Mauris nec fringilla odio. Sed vestibulum purus ut imperdiet rutrum.<br />
+				<br /><br />
+				<section class=h4><h4 id="With-its-own-H4">With its own H4</h4><br />
+					Curabitur vitae lacus commodo, suscipit dolor tincidunt, tempor orci. Morbi quis erat ligula. Sed arcu eros, sodales ac urna eget, facilisis lobortis tortor. Morbi felis metus, tincidunt eu gravida et, scelerisque dapibus augue. Fusce aliquam condimentum euismod. Orci varius natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Maecenas nec porta mi. Curabitur est magna, ultricies non congue id, condimentum sit amet magna. Aenean sagittis purus massa, condimentum gravida lorem auctor eget. <br />
+					<br /><br />
+				</section><br />
+			</section><br />
+		</section><br />
+
+</article>    </body>
+</html>
