+++ main.py
@@ -1,5 +1,9 @@
 from private import base_path, templates
 
+import os
+
+from chunk_markdown import chunk_markdown_file, print_nested
+
 from handlers.frontmatter import handle_frontmatter
 from handlers.callout import handle_callouts
 from handlers.merge_with_post import merge_with_post
@@ -8,95 +12,6 @@ from handlers.links import handle_links
 from handlers.preformatted import handle_preformatted_blocks
 from handlers.tables import handle_tables
 from handlers.regex import convert_formatting
-
-def get_rough_blocks_of_content(filepath: str) -> list[str]:
-
-    with open(filepath, "r") as file:
-        lines = file.readlines()
-
-    last_line = len(lines) -1
-        
-    list_by_blocks_of_content = [ ]
-    content_block = [ ]
-    block_started = False
-    for i, line in enumerate(lines):
-        if line == "\n":
-            if block_started:
-                block_started = False
-                list_by_blocks_of_content.append(content_block)
-                
-            if not block_started:
-                block_started = True
-                content_block = [ ]
-        else:
-            if block_started:
-                content_block.append(line.strip('\n'))
-        if i == last_line:
-            list_by_blocks_of_content.append(content_block)
-
-    return list_by_blocks_of_content
-
-def group_blocks_by_header(rough_blocks: list[str]):
-    headers = ['######', '#####', '####', '###', '##', '#']
-    headers.reverse()
-    grouped_by_headers = { }
-    
-    for header in headers:
-        block_key = ""
-        for block in rough_blocks:
-            if header in block[0].split(" ")[0]:
-                block_key = block[0]
-                grouped_by_headers[block_key] = [ ]
-            elif block_key:
-                if block[0].split(" ")[0] not in headers:
-                    grouped_by_headers[block_key].append(block)
-                else:
-                    block_key = ""
-    
-    return grouped_by_headers
-
-def nest_subheadings_inside_headings(grouped_by_headers):
-    headers = ['######', '#####', '####', '###', '##']
-    headers.reverse()
-    
-    by_h2 = { }
-    last_header = { }
-    
-    for header_key in list(grouped_by_headers.keys()):
-        cleaned_key = header_key.split(" ")[0]
-        last_header[cleaned_key] = header_key
-        
-        match cleaned_key:
-            case "##":
-                by_h2[header_key] = {"content" : grouped_by_headers[header_key]}
-            case "###":
-                by_h2[last_header["##"]][header_key] = {"content" : grouped_by_headers[header_key]}
-            case "####":
-                by_h2[last_header["##"]][last_header["###"]][header_key] = {"content" : grouped_by_headers[header_key]}
-            case "#####":
-                by_h2[last_header["##"]][last_header["###"]][last_header["####"]][header_key] = {"content" : grouped_by_headers[header_key]}
-            case _:
-                by_h2[last_header["##"]][last_header["###"]][last_header["####"]][last_header["#####"]][header_key] = {"content" : grouped_by_headers[header_key]} 
-    
-    return by_h2
-
-def print_nested(by_h2):
-    for h2, h2_content in by_h2.items():
-        print(h2)
-        for h3, h3_content in h2_content.items():
-            print("  " + h3)
-            if isinstance(h3_content, dict):
-                for h4, h_4content in h3_content.items():
-                    print("    "+h4)
-                    print(f"      {h_4content}")
-            else:
-                for content in h3_content:
-                    print(f"    {content}")
-
-rough_blocks = get_rough_blocks_of_content(f"{base_path}\\raw\\example.md")
-grouped_by_headers = group_blocks_by_header(rough_blocks)
-nested_content = nest_subheadings_inside_headings(grouped_by_headers)
-# print_nested(nested_content)
 
 def format_content(nested_blocks_of_content):
     html_content = ""
@@ -127,7 +42,9 @@ def format_content(nested_blocks_of_content):
                                     else:
                                         h6_prefix, h6_indent, h6_suffix = get_block_wrapper(h6)
                                         h6_block = ""
-                                    h5_block += h6_prefix + h6_block + h6_suffix
+                                        for sub_h6, sub_h6_content in h6_content.items():
+                                            h6_block += format_block(h6_indent, sub_h6_content)
+                                        h5_block += h6_prefix + h6_block + h6_suffix
                                 h4_block += h5_prefix + h5_block + h5_suffix
                         h3_block += h4_prefix + h4_block + h4_suffix
                 h2_block += h3_prefix + h3_block + h3_suffix
@@ -141,7 +58,6 @@ def get_block_wrapper(header):
     indent_level = ""
     for i in range(0, header_level):
         indent_level += "\t"
-    indent_level = ""
     header = header.strip("#").strip(" ")
     header_id = "-".join(header.split(" "))
     block_prefix = f'{indent_level}<section class=h{header_level}><h{header_level} id="{header_id}">{header}</h{header_level}>\n'
@@ -153,7 +69,7 @@ def format_block(indent_level, block_of_content):
     formatted_block = ""
     block = ""
     indent_level += "\t"
-    indent_level = ""
+
     for sub_block in block_of_content:
         sub_block = handle_tables(sub_block)
         sub_block = handle_callouts(sub_block)
@@ -169,26 +85,101 @@ def format_block(indent_level, block_of_content):
     formatted_block = block 
     return formatted_block
 
-html_string = format_content(nested_content)
 
-title = "raw"
-
-# with open(f"{base_path}\\raw.html", "w+") as file:
-#     file.write(html_string)
+def load_files(source_directory: str, destination_directory: str, fresh_build:bool = False):
+    files_to_convert = { }
+    files_in_source = os.listdir(source_directory)
     
-base_fp = f"{base_path}{templates["base"]}"
-base_html_template = [ ]
+    if fresh_build:
+        for file in files_in_source:
+            if file.endswith(".md"):
+                with open(source_directory + file, "r") as reader:
+                    files_to_convert[file] = reader.readlines()
+        return files_to_convert
+    
+    # Not a fresh build, we only want to convert what has changed
+    files_in_destination = os.listdir(destination_directory)
+    for src_file in files_in_source:
+        if not src_file.endswith(".md"): #skip anything that's not markdown
+            continue
+        for dest_file in files_in_destination:
+            if not dest_file.endswith(".html"): #skip anything thats not html
+                continue
+            if src_file.strip(".md").lower() == dest_file.strip(".html").lower():
+                src_last_modified = os.path.getmtime(source_directory + src_file)
+                dest_last_modified = os.path.getmtime(destination_directory + dest_file)
+                if src_last_modified <= dest_last_modified:
+                    continue
+        with open(source_directory + src_file, "r") as reader:
+            files_to_convert[src_file] = reader.readlines()
+    return files_to_convert
 
-with open(base_fp, "r") as file:
-    base_html_template = file.readlines()
 
-for i, line in enumerate(base_html_template):
-    if "{{title}}" in line:
-        base_html_template[i] = line.replace("{{title}}", title)
-    if "{{body}}" in line:
-        base_html_template[i] = line.replace("{{body}}", html_string)
+def old_core_loop(nested_content):
+    html_string = format_content(nested_content)
 
-new_html = f"{base_path}\\content\\{title}.html"
+    title = "raw"
 
-with open(new_html, "w+") as file:
-    file.writelines(base_html_template)
+    # with open(f"{base_path}\\raw.html", "w+") as file:
+    #     file.write(html_string)
+        
+    base_fp = f"{base_path}{templates["base"]}"
+    base_html_template = [ ]
+
+    with open(base_fp, "r") as file:
+        base_html_template = file.readlines()
+
+    for i, line in enumerate(base_html_template):
+        if "{{title}}" in line:
+            base_html_template[i] = line.replace("{{title}}", title)
+        if "{{body}}" in line:
+            base_html_template[i] = line.replace("{{body}}", html_string)
+
+    new_html = f"{base_path}\\content\\{title}.html"
+
+    with open(new_html, "w+") as file:
+        file.writelines(base_html_template)
+
+
+def main():
+    files = load_files(f"{base_path}\\raw\\", f"{base_path}\\content\\", False)
+
+    for fname, content in files.items():
+            # Get frontmatter
+        title, date_created, dates_modified, markdown_body = handle_frontmatter(content)
+        frontmatter = (title, date_created, dates_modified[-1])
+        chunked_content = chunk_markdown_file(markdown_body)
+        files[fname] = {
+            "frontmatter" : frontmatter,
+            "content"     : chunked_content
+        }
+    # Handle embedded stuff
+    
+    # Finish conversion and make output file
+    for fname in list(files.keys()):
+        frontmatter = files[fname]["frontmatter"]
+        body_content = files[fname]["content"]
+        
+        formatted_content = format_content(body_content)
+        
+        squished_post = merge_with_post(formatted_content, frontmatter)
+        
+        base_fp = f"{base_path}{templates["base"]}"
+        base_html_template = [ ]
+
+        with open(base_fp, "r") as file:
+            base_html_template = file.readlines()
+
+        for i, line in enumerate(base_html_template):
+            if "{{title}}" in line:
+                base_html_template[i] = line.replace("{{title}}", frontmatter[0])
+            if "{{body}}" in line:
+                base_html_template[i] = line.replace("{{body}}", squished_post)
+
+        new_html = f"{base_path}\\content\\{frontmatter[0]}.html"
+
+        with open(new_html, "w+") as file:
+            file.writelines(base_html_template)
+        
+if __name__ == "__main__":
+    main()
