+++ raw/embedded_example.md
@@ -5,6 +5,8 @@ modified:
     - "2025-01-11"
 ---
 
+## Content from Example
+
 This is an example of embedded content:
 
 ![[example#Now onto some code]]
